requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
python-dotenv==1.0.1
langchain==0.3.14
langchain-groq==0.2.4
langgraph==0.2.65
